add: New option to specify binary to play mp3-files
For Linux the default player for mp3 is omxplayer which is not installed on headless raspbian buster. Installing it via apt-get would trigger installation of a graphics stack (mesa-va-drivers mesa-vdpau-drivers omxplayer va-driver-all etc.). To omit that now the user can specify an alternative mp3 player (e.g. mpg321). Defaults are as before.